Update web-publish-options.md
> and each one will maintain its assigned IP address as long as desired.
How is desire related to IP not changing? When does it change?
Any chance you could add a comment that App Service is much simpler to get a web app set up? Maybe that's all in the comparison.

Long sentences are difficult for me and impossible for ESL and our machine translators.