Update build-tools-container-caveats.md
The Microsoft Writers Style Guide recommends that we use a numbered list when actions much occur in a sequence, as in a procedure, or when items in a list must be presented in priority order. I could be wrong, but I don't think either of those two cases are true for the content on this page.