Update finding-and-using-visual-studio-extensions.md
This is to add information related to the new Extension UI-Delay notifications (externally "Unresponsiveness Notifications"), which is similar to Extension Crash Notifications but shows notifications for extension causing VS UI to be unresponsive. Michael (author of Crash notifications and I) wanted to put these 2 under one node because of the shared functionality as well as similarity in nature.