New CodeLens test indicator behavior
This is an intentional change we have made to code lens test indicators to improve performance. The test indicators need the source information (where the test is in the file) to determine where to draw the indicators and fetching source information is expensive. So we have made it lazy. It will only happen when you build the solution or click on the test to navigate to source. This has resulted in significant performance improvements.