Update how-to-add-class-diagrams-to-projects.md
Updated some text for clarity.

I also added something that I think is needed, but maybe not everywhere:

When referring to a window, I see an approach that is succinct, but not always clear. For ex: "In **Solution Explorer**...", and "In **Class View**...". It's short and sweet but not clear if you don't already know VS. The solution explorer can probably be left as is, but saying, "In Class View", I think, is really vague. So I added some text so it reads, "In the **Class View** window, so it's clearly not a class file but a window. It might not be needed for the solution explorer but it's clearer there too. [Of course, that could mean updating everywhere...]