C32669: Changing Italis asterisk to underscore to avoid localization issues
Hello, @mikejo5000,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
"several asterisk are breaking the code in localized pages: Example:
Eliminare \Users\ ***{nome}* \AppData\Local\Microsoft\VisualStudio\*Exp
at this page: https://docs.microsoft.com/it-IT/visualstudio/modeling/how-to-change-the-namespace-of-a-domain-specific-language#to-change-the-namespace-of-a-domain-specific-language"
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.