V2515: Broken link
Hello @TerryGLee,
there is a broken link in the page in this string:
For a related video demonstration, see How do I: Automate PowerPoint from an Excel Add-in?.
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.