Clarification around snaphot capture
We overused the the term "snapshot" to mean multiple things as both a noun and verb.

Our "Process Snapshot" feature is not supported in the IntelliTrace standalone tool so we are removing the term "snaphot" to reduce confusion.

Additionally when we start the discussion about collecting snapshots on the main IntelliTrace page we are explicitly stating that the stand-alone collection does not support this type of snapshot.