Remove 'stopping event' from Load Complete Event
The documentation for IDebugLoadCompleteEvent2 indicated that load complete must be a stopping event. While it is allowed for load complete to be a stopping event, there isn't really a reason to do so aside from that fact that this is one way that a debug engine author might want to make sure that they don't continue execution until after the initial set of breakpoints/exception settings are received from the UI. But there are other approaches to this as well, and none of the recent Microsoft-implemented debug engines have done things that way. So this removes the remark.