C87692: Escaping mock URL to
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
Description:   As this is a mock hyperlink not a real one, it should not be rendered
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.
@mikejo5000