Fix for Visual Studio for Mac Workloads link
The current link resolves to a 404 error:
https://github.com/MicrosoftDocs/visualstudio-docs/blob/master/visualstudio/mac/workloads/

The correct link needs to be:
https://github.com/MicrosoftDocs/visualstudio-docs/blob/master/mac/workloads.md

Notice that after "master" there is no "visualstudio" folder in the Mac Visual Studio repo, so it just needs to be "mac". Unless of course the whole thing needs to be moved to a "visualstudio" folder, but that's another matter on its own.