Update container-launch-settings.md
The documentation for the `launchUrl` settings were not rendering correctly. This may be because they were enclosed in chevrons ("<" and ">") -- I'm not sure. However, in VS2019, the `launchSettings.json` file has these contained within curly braces:
```
    "Docker": {
      "commandName": "Docker",
      "launchBrowser": true,
      "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}",
      "environmentVariables": {
        "ASPNETCORE_URLS": "https://+:443;http://+:80",
        "ASPNETCORE_HTTPS_PORT": "44391"
      },
      // etc.
```
I have changed this document so that the rendering of the substitutions appears to be correct, and I have corrected the casing of the substitutions to be how they appear with the `launchSettings.json` file.