Update how-to-save-and-open-files-with-encoding.md
Add note about requiring project files to be unloaded before the **Save File As** option is available for them. I have assumed here the reader knows how to unload a project in Visual Studio, though it may be helpful to link to relevant docs which explain that (if they exist).