Update info on MiddleLayer in LSP Extension docs (#6831)
As of Visual Studio 2019, the documentation on adding a Middle Layer to intercept Language Server Protocol messages is outdated. It causes extensions following the example in the docs to never call their language server (as mentioned in this issue: https://github.com/MicrosoftDocs/visualstudio-docs/issues/4735).
This PR updates the relevant references/examples on implementing a "Middle Layer" in VS2019.