Update xmlpeek-task.md
Every single time I try to use XPath on an XML file it doesn't work.  This is because all the XML files I've ever needed to use it on just happen to make use of a default namespace.  There is a caveat that needs to be understood when the default namespace is being used, which this article does not cover.  It would be nice if the examples were updated to include it explicitly, but in the very least, this additional "See also" link should be referenced.  It wasn't until I stumbled upon this other article that I finally figured out how to make XPath work properly.