Fix MD link to image in vs-ide-benefit.md (#7486)
On the [The Visual Studio IDE](https://docs.microsoft.com/en-us/visualstudio/subscriptions/vs-ide-benefit) page, in Activation Step 2, there is a double quote that is closed prematurely (after the word 'Download'), which causes the image to not be shown correctly. The fix is to enclose the word 'Download' in single quotes, similar to what is done in Step 5.