Update visual-studio-sdk-glossary.md (#7429)
* Update visual-studio-sdk-glossary.md

I came across this glossary which contains very useful information, but it was borderline illegible because all the terms and definitions were joined together without any deliniation between where a term ends and its definition starts.  Took a crack at repairing the layout.  It looks like each term should really be on its own line as well, but I'm not sure why they aren't rendering like that.  They are on their own line in the editor.

* PR review edit, two trailing spaces for single line break

Co-authored-by: Kristine Toliver <ktoliver@users.noreply.github.com>