Remove second and from list of projects types
The sentence is easier to read without the second "and."  Each project type and Azure functions should just be listed one after the other for clarity.

Note: as a non-Microsoft reader, the current branch of this documentation is the first I learned that VS2022 includes Docker support for Azure functions. The change was approved so it must be true.