Project type and targets were used interchangeably
For VS2017: the sentence was:
Supported project types include .NET Framework and .NET Core web and console apps.

This confused the project type (type web or type console) with the targetted framework (.NET Core/.NET Framework). Using "apps" is confusing as the project templates when selected Visual Studio are App/Application singular (depending on the version) To make this clearer app was used and not apps and the project type was separated from the project target:

Supported project types include web app and console app targeting .NET Framework and .NET Core.

For >=VS2019 the issue was identical to VS2017 except Azure Function was also a project type:

Original:
Supported project types include .NET Framework, .NET Core web and console apps, and Azure Functions.

Updated:
Supported project types include web app, console app, and Azure Function targeting .NET Framework and .NET Core.

The second sentence VS2107 was
In this article, we use ASP.NET Core web apps and .NET Framework console apps.

VS2017 uses the name application the project templates listed in Visual Studio not app and the use of the pronoun "we" was removed -- it is an article singular, not some plural entity :
The examples presented in this article, are an ASP.NET Core web application and a .NET Framework console application.

For >=VS2019 the original sentence was:
In this article, we use ASP.NET Core web apps and .NET Framework console apps.

The above sentence is confusing b/c it uses apps (plural) and there is only a single ASP.NET Core Web App presented and a single Console App (.NET Framework) presented. I used the text associated with the project template demonstrated in the article "ASP.NET Core Web App" and  "Console App (.NET Framework)"):
The examples presented in this article, are a project of type ASP.NET Core Web App and a project of type Console App (.NET Framework).