Justifying container orchestration
The original intro to this section read:
When you use .NET Framework console app projects, the option to add Docker support without orchestration isn't supported. You can still use the following procedure, even if you're using only a single Docker project.

There text to guide the reader as to what is orchestration and how it relates to debugging a .NET Framework console app in a local Docker container. The reader is not told that if they add orchestration, it also enables Docker support for the project which allows the project to be debugged in a local Docker container (the title of this help topic).

The other thing that should be reinforced is that this approach applies to .NET Framework console app projects. VS2022 supports adding Docker support to .NET Core console app project for example. Adding docker support by adding support for orchestration applies to a very small number scenarios.