Indenting link descriptions
Without proper indentation, each lined seemed like to be describing the following link and not the preceding one.  Adding two spaces of whitespace at the end of the bulleted line (which inserts a newline instead of soft wrapping) and four spaces in front of the description (for the indentation) fixes this issue.