Remove VS2019 steps
Taking VS2019 out of this tutorial except some notes about how to try to set it up yourself. The experience is not good with the current versions of .NET and WASDK templates out there. I've saved a copy with these 2019 steps, but so far I can't get it set up to take screenshots.