Update supported edits doc to point to Roslyn
Point to the Roslyn wiki https://github.com/dotnet/roslyn/blob/main/docs/wiki/EnC-Supported-Edits.md
for the edit types supported and the list of proposed future enhancements.
This list is the most up to date. And sometimes folks wonder what's
coming and want to vote on potential enhancements, so visibility
into that is good as well.