Disambiguate from-scratch project from csproj
MSBuild doesn't usually have very strong coupling to file extension, but some things (parts of the .NET SDK, parts of Visual Studio, and lots of third-party code) will assume that a file named `*.csproj` will walk and quack like a "real"/"standard" `.csproj`.

Since that's not the case for the built-from-scratch-as-an-exercise project in this walkthrough, let's name it something else.