Update publish-page-project-designer.md
I am one of the Security Program Manager at MSRC Microsoft team. We have an external vulnerability submission which highlights that there is human error in this article and the value here must be 65535 instead of 65335
https://learn.microsoft.com/en-us/visualstudio/ide/reference/publish-page-project-designer?view=vs-2019#:~:text=maximum%20value%20of-,65355,-(MaxValue)%2C%20the 

The same was confirmed by the Vulnerability and Assessment team to update the article. Hence, requesting to update the change